WheelSense-Server/requirements.txt
# WheelSense Server Requirements (Python)

# FastAPI and dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# MQTT Client
paho-mqtt==1.6.1

# HTTP Requests
requests==2.31.0

# Database
# Note: SQLite is built-in to Python, no need to install

# Optional: PostgreSQL support
# psycopg2-binary==2.9.9
# sqlalchemy==2.0.23

# Development tools
python-dotenv==1.0.0






